mcp/distribute-metal-mcp/pyproject.toml
[project]
name = "distribute-metal-mcp"
version = "0.1.0"
description = "MCP server for DistributeMetal — cluster inspection, YAML generation, and validation"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
distribute-metal-mcp = "distribute_metal_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/distribute_metal_mcp"]
